renamed to interfaceutils moved findOutgoingPcapInterfaceNameAndIp
