+++ interfaceutils.go
@@ -1,0 +1,57 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"net"
+
+	"github.com/google/gopacket/pcap"
+)
+
+func FindPcapInterfaceName(ipAddress net.IP) (string, error) {
+	devices, err := pcap.FindAllDevs()
+	if err != nil {
+		return "", err
+	}
+
+	for _, networkInterface := range devices {
+		interfaceAddresses := networkInterface.Addresses
+		for _, interfaceAddress := range interfaceAddresses {
+			interfaceIp := interfaceAddress.IP
+			if interfaceIp.Equal(ipAddress) {
+				return networkInterface.Name, nil
+			}
+		}
+	}
+
+	return "", nil
+}
+
+func findOutgoingPcapInterfaceNameAndIp(targetIp *net.IPAddr) (string, *net.IPAddr, error) {
+	initialConn, err := net.DialUDP("udp", nil, &net.UDPAddr{IP: targetIp.IP, Port: 443})
+	if err != nil {
+		return "", nil, err
+	}
+
+	localInterfaceIp := initialConn.LocalAddr().(*net.UDPAddr).IP
+	_ = initialConn.Close()
+
+	outgoingPcapInterfaceName, err := FindPcapInterfaceName(localInterfaceIp)
+	if err != nil {
+		return "", nil, err
+	}
+	if outgoingPcapInterfaceName == "" {
+		return "", nil, errors.New(
+			fmt.Sprintf("Unable to lookup the outgoing interface for local IP: %s", localInterfaceIp))
+	}
+
+	_, localNet, _ := net.ParseCIDR("127.0.0.0/8")
+	if localNet.Contains(localInterfaceIp) {
+		return "", nil, errors.New(
+			"Outgoing interface is local. Either the destination is the local machine or a" +
+				" local proxy is being used.\nPlease choose a remote destination or exclude this app from being" +
+				" proxied and try again.")
+	}
+
+	return outgoingPcapInterfaceName, &net.IPAddr{IP: localInterfaceIp}, nil
+}
